Add invalid credentials test
