+++ client/src/app/components/userAuth/new.html
@@ -2,6 +2,10 @@
   <acme-navbar></acme-navbar>
 </header>
 
+<div class="alert alert-danger" role="alert" ng-show="error">
+  {{ error }}
+</div>
+
 <form ng-submit="submitLogin(loginForm)" role="form" ng-init="loginForm = {}">
   <div class="form-group">
     <label for="email">Email</label>
