+++ client/src/app/components/userAuth/signIn.controller.js
@@ -1,0 +1,9 @@
+export class SignInController {
+  constructor ($scope) {
+    'ngInject';
+
+    $scope.$on('auth:login-error', function(ev, reason) {
+      $scope.error = reason.errors[0];
+    });
+  }
+}
