--- client/src/app/components/userAuth/sign_in.controller.js
@@ -1,5 +1,0 @@
-export class SignInController {
-  constructor ($scope) {
-    'ngInject';
-  }
-}
