--- spec/support/pages/login_page.rb
@@ -1,9 +1,0 @@
-class LoginPage
-  include Capybara::DSL
-
-  def sign_in(email, password)
-    fill_in "Email", with: email
-    fill_in "Password", with: password
-    find("button", text: "Sign in").click
-  end
-end
